fix migration

+++ database/migrations/2018_11_06_193918_add_locked_field_to_threads_table.php
@@ -26,7 +26,9 @@ class AddLockedFieldToThreadsTable extends Migration
     public function down()
     {
         Schema::table('threads', function (Blueprint $table) {
-            $table->dropColumn('locked');
+            if (Schema::hasColumn('threads', 'locked')) {
+                $table->dropColumn('locked');
+            }
         });
     }
 }
